opendal_core::blocking

Struct StdReader

pub struct StdReader { /* private fields */ }

Available on crate feature blocking only.

StdReader is the adapter of Read, Seek and BufRead for [BlockingReader][crate::BlockingReader].

Users can use this adapter in cases where they need to use Read or BufRead trait.

StdReader also implements Send and Sync.