opendal_core::blocking

Struct StdWriter

pub struct StdWriter { /* private fields */ }

Available on crate feature blocking only.

StdWriter is the adapter of std::io::Write for [BlockingWriter].

Users can use this adapter in cases where they need to use std::io::Write related trait.

Notes

Files are automatically closed when they go out of scope. Errors detected on closing are ignored by the implementation of Drop. Use the method close if these errors must be manually handled.

Implementations

impl StdWriter

pub fn close(&mut self) -> Result<()>

Close the internal writer and make sure all data have been stored.