opendal_core::options

Struct ReaderOptions

pub struct ReaderOptions {
    pub version: Option<String>,
    pub if_match: Option<String>,
    pub if_none_match: Option<String>,
    pub if_modified_since: Option<Timestamp>,
    pub if_unmodified_since: Option<Timestamp>,
    pub concurrent: usize,
    pub chunk: Option<usize>,
    pub gap: Option<usize>,
    pub prefetch: usize,
}

Options for reader operations.

Fields
version: Option<String>

Set version for this operation.

This option can be used to retrieve the data of a specified version of the given path.

If the version doesnโt exist, an error with kind [ErrorKind::NotFound] will be returned.

if_match: Option<String>

Set if_match for this operation.

This option can be used to check if the fileโs ETag matches the given ETag.

If file exists and itโs etag doesnโt match, an error with kind [ErrorKind::ConditionNotMatch] will be returned.

if_none_match: Option<String>

Set if_none_match for this operation.

This option can be used to check if the fileโs ETag doesnโt match the given ETag.

If file exists and itโs etag match, an error with kind [ErrorKind::ConditionNotMatch] will be returned.

if_modified_since: Option<Timestamp>

Set if_modified_since for this operation.

This option can be used to check if the file has been modified since the given timestamp.

If file exists and it hasnโt been modified since the specified time, an error with kind [ErrorKind::ConditionNotMatch] will be returned.

if_unmodified_since: Option<Timestamp>

Set if_unmodified_since for this operation.

This feature can be used to check if the file hasnโt been modified since the given timestamp.

If file exists and it has been modified since the specified time, an error with kind [ErrorKind::ConditionNotMatch] will be returned.

concurrent: usize

Set concurrent for the operation.

OpenDAL by default to read file without concurrent. This is not efficient for cases when users read large chunks of data. By setting concurrent, opendal will reading files concurrently on support storage services.

By setting concurrent, opendal will fetch chunks concurrently with the give chunk size.

Refer to crate::docs::performance for more details.

chunk: Option<usize>

Set chunk for the operation.

OpenDAL will use servicesโ preferred chunk size by default. Users can set chunk based on their own needs.

Refer to crate::docs::performance for more details.

gap: Option<usize>

Controls the optimization strategy for range reads in [Reader::fetch].

When performing range reads, if the gap between two requested ranges is smaller than the configured gap size, OpenDAL will merge these ranges into a single read request and discard the unrequested data in between. This helps reduce the number of API calls to remote storage services.

This optimization is particularly useful when performing multiple small range reads that are close to each other, as it reduces the overhead of multiple network requests at the cost of transferring some additional data.

Refer to crate::docs::performance for more details.

prefetch: usize

Controls the number of prefetched bytes ranges that can be buffered in memory during concurrent reading.

When performing concurrent reads with Reader, this option limits how many completed-but-not-yet-read chunks can be buffered. Once the number of buffered chunks reaches this limit, no new read tasks will be spawned until some of the buffered chunks are consumed.

Default value: 0 (no prefetching, strict back-pressure control)

Set to a higher value to allow more aggressive prefetching at the cost of memory

This option helps prevent memory exhaustion when reading large files with high concurrency settings.