opendal_core

Struct OperatorBuilder

pub struct OperatorBuilder<A: Access> { /* private fields */ }

OperatorBuilder is a typed builder to build an Operator.

Notes

OpenDAL uses static dispatch internally and only performs dynamic dispatch at the outmost type erase layer. OperatorBuilder is the only public API provided by OpenDAL come with generic parameters.

It’s required to call finish after the operator built.

Examples

For users who want to support many services, we can build a helper function like the following:

use std::collections::HashMap;

use opendal_core::services;
use opendal_core::Builder;
use opendal_core::Operator;
use opendal_core::Result;

fn init_service<B: Builder>(cfg: HashMap<String, String>) -> Result<Operator> {
    let op = Operator::from_iter::<B>(cfg)?
        // add layers
        // .layer(LoggingLayer::default())
        // .layer(RetryLayer::new())
        .finish();

    Ok(op)
}

async fn init(scheme: &str, cfg: HashMap<String, String>) -> Result<()> {
    let _ = match scheme {
        services::MEMORY_SCHEME => init_service::<services::Memory>(cfg)?,
        _ => todo!(),
    };

    Ok(())
}

Implementations

impl<A: Access> OperatorBuilder<A>

pub fn new(accessor: A) -> OperatorBuilder<impl Access>

Create a new operator builder.

pub fn layer<L: Layer<A>>(self, layer: L) -> OperatorBuilder<L::LayeredAccess>

Create a new layer with static dispatch.

Notes

OperatorBuilder::layer() is using static dispatch which is zero cost. Operator::layer() is using dynamic dispatch which has a bit runtime overhead with an extra vtable lookup and unable to inline.

It’s always recommended to use OperatorBuilder::layer() instead.

Examples

let client = HttpClient::new()?;
let op = Operator::new(Memory::default())?
    .layer(HttpClientLayer::new(client))
    .finish();
// All operations will go through the new_layer
let _ = op.read("test_file").await?;

pub fn finish(self) -> Operator

Finish the building to construct an Operator.