opendal_core

Module services

Services will provide builders to build underlying backends.

More ongoing services support is tracked at opendal#5. Please feel free to submit issues if there are services not covered.

Structs

Gdrive

GoogleDrive backend support.

GdriveConfig

GoogleDrive configuration.

Gridfs

Capabilities

GridfsConfig

Config for Grid file system support.

Hdfs

A distributed file system that provides high-throughput access to application data.

HdfsConfig

Hadoop Distributed File System (HDFS™) support.

Ipmfs

IPFS file system support based on IPFS MFS API.

IpmfsConfig

Config for IPFS MFS support.

Lakefs

Lakefs’s API support. This service will visit the Lakefs API to access the Lakefs File System. Currently, we only support the model and dataset types of repositories, and operations are limited to reading and listing/stating.

LakefsConfig

Configuration for Lakefs service support.

Memory

In memory service support. (BTreeMap Based)

MemoryConfig

Config for memory.

Mongodb

Capabilities

MongodbConfig

Config for Mongodb service support.

Monoiofs

File system support via [monoio].

MonoiofsConfig

Config for monoiofs services support.

Redis

Redis services support.

RedisConfig

Config for Redis services support.

Rocksdb

RocksDB service support.

RocksdbConfig

Config for Rocksdb Service.

Sftp

SFTP services support. (only works on unix)

SftpConfig

Config for Sftp Service support.

VercelArtifacts

Vercel Cache backend support.

VercelArtifactsConfig

Config for Vercel Cache support.

Webdav

WebDAV backend support.

WebdavConfig

Config for WebDAV backend support.

Webhdfs

WebHDFS’s REST API support. There two implementations of WebHDFS REST API:

WebhdfsConfig

Config for WebHDFS support.

YandexDisk

YandexDisk services support.

YandexDiskConfig

Config for YandexDisk services support.

Constants

GDRIVE_SCHEMEservices-gdrive

Default scheme for gdrive service.

GRIDFS_SCHEMEservices-gridfs

Default scheme for gridfs service.

HDFS_SCHEMEservices-hdfs

Default scheme for hdfs service.

IPMFS_SCHEMEservices-ipmfs

Default scheme for ipmfs service.

LAKEFS_SCHEMEservices-lakefs

Default scheme for lakefs service.

MEMORY_SCHEMEservices-memory

Default scheme for memory service.

MONGODB_SCHEMEservices-mongodb

Default scheme for mongodb service.

MONOIOFS_SCHEMEservices-monoiofs

Default scheme for monoiofs service.

REDIS_SCHEMEservices-redis

Default scheme for redis service.

ROCKSDB_SCHEMEservices-rocksdb

Default scheme for rocksdb service.

SFTP_SCHEMEservices-sftp

Default scheme for sftp service.

VERCEL_ARTIFACTS_SCHEMEservices-vercel-artifacts

Default scheme for vercel-artifacts service.

WEBDAV_SCHEMEservices-webdav

Default scheme for webdav service.

WEBHDFS_SCHEMEservices-webhdfs

Default scheme for webhdfs service.

YANDEX_DISK_SCHEMEservices-yandex-disk

Default scheme for yandex-disk service.